+++ assets/images/index.js
@@ -1,0 +1,4 @@
+export default {
+  graph: require('./graph.png'),
+  splash: require('./splash.png'),
+};
